chore: quick fixes
